# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "API of sowgatly.app",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        },
        {
            "url": "https://sowgatly.app",
            "description": "Remote server"
        }
    ],
    "paths": {
...
        "/api/products/{id}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Reviews of a product (public). Send `page` for a paginated list.",
                "operationId": "83d3abb23ec8d3f8ba035b73ec250361",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reviews with rating_match / rating_value / rating_service, author and meta (count, average, can_review, my_review)"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Review a product you have ordered (one review per product, updates replace it)",
                "operationId": "355b3f3f624ae313c3587c942ab160b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rating": {
                                        "description": "Optional when the three criteria are given; then it is their rounded average",
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "rating_match": {
                                        "description": "Соответствие",
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "rating_value": {
                                        "description": "Цена / качество",
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "rating_service": {
                                        "description": "Сервис магазина",
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comment": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "nullable": true
                                    },
                                    "order_id": {
                                        "description": "One of the caller's orders containing the product",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Saved review"
                    },
                    "403": {
                        "description": "Not a buyer of this product / order mismatch"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/shop/orders": {
            "get": {
                "tags": [
                    "Shop orders"
                ],
                "summary": "Orders placed with the caller's own shop",
                "operationId": "a0f36b66b633c58586fbe2f004e802d4",
                "responses": {
                    "200": {
                        "description": "Orders, newest first"
                    },
                    "403": {
                        "description": "The caller has no shop"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shop/orders/{id}/status": {
            "put": {
                "tags": [
                    "Shop orders"
                ],
                "summary": "Move one of the shop's orders to the next status",
                "operationId": "3c8d14b94937e7154a3d2c0871e586c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "processing",
                                            "delivering",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated order"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "422": {
                        "description": "Transition not allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}